changed int to size_t

+++ 16-binary_tree_is_perfect.c
@@ -38,9 +38,9 @@ int binary_tree_is_perfect(const binary_tree_t *tree)
  *
  * Return: Height of the tree, or 0 if tree is NULL.
  */
-int binary_tree_height(const binary_tree_t *tree)
+size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int left_height, right_height;
+	size_t left_height, right_height;
 
 	/* Check if tree is NULL */
 	if (tree == NULL)
